move chunked decoder and add tests
